+++ node/clients/Organizations.ts
@@ -7,7 +7,7 @@ import { getTokenToHeader } from './index'
 
 export class OrganizationsGraphQLClient extends AppGraphQLClient {
   constructor(ctx: IOContext, options?: InstanceOptions) {
-    super('vtex.b2b-organizations-graphql@0.x', ctx, options)
+    super('vtex.b2b-organizations-graphql@1.x', ctx, options)
   }
 
   public getAddresses = async (costCenterId: string): Promise<any> => {
@@ -15,8 +15,8 @@ export class OrganizationsGraphQLClient extends AppGraphQLClient {
       {
         extensions: {
           persistedQuery: {
-            provider: 'vtex.b2b-organizations-graphql@0.x',
-            sender: 'vtex.b2b-checkout-settings@1.x',
+            provider: 'vtex.b2b-organizations-graphql@1.x',
+            sender: 'vtex.b2b-checkout-settings@2.x',
           },
         },
         query: QUERIES.getAddresses,
@@ -25,8 +25,8 @@ export class OrganizationsGraphQLClient extends AppGraphQLClient {
         },
       },
       {
+        headers: getTokenToHeader(this.context),
         params: {
-          headers: getTokenToHeader(this.context),
           locale: this.context.locale,
         },
       }
@@ -38,8 +38,8 @@ export class OrganizationsGraphQLClient extends AppGraphQLClient {
       {
         extensions: {
           persistedQuery: {
-            provider: 'vtex.b2b-organizations-graphql@0.x',
-            sender: 'vtex.b2b-checkout-settings@1.x',
+            provider: 'vtex.b2b-organizations-graphql@1.x',
+            sender: 'vtex.b2b-checkout-settings@2.x',
           },
         },
         query: QUERIES.getOrganizationDetails,
@@ -48,8 +48,8 @@ export class OrganizationsGraphQLClient extends AppGraphQLClient {
         },
       },
       {
+        headers: getTokenToHeader(this.context),
         params: {
-          headers: getTokenToHeader(this.context),
           locale: this.context.locale,
         },
       }